requirements.txt
requests>=2.32.3
python-dotenv>=1.0.1
flask>=3.0.0
fastapi>=0.115.0
openai>=1.50.0
tiktoken>=0.7.0
model-context-protocol>=0.4.0
redis>=5.0.0

